gpio: Fix JS binding regarding interrupt injections
According to [1] and based on stress tests, it is not correct to call
uv_queue_work outside of the loop thread. We rather need to use the
async API of libuv. That even simplifies things.

Resolves "uv__queue_done: Assertion `uv__has_active_reqs(req->loop)' failed"
errors that were easy to trigger by multiple DIs being used in parallel.
See also [2].

[1] https://github.com/libuv/libuv/discussions/3847
[2] https://github.com/siemens/meta-iot2050/issues/386

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>